empty message

+++ src/test/java/com/jiker/keju/TaxPaymentTest.java
@@ -16,19 +16,19 @@ public class TaxPaymentTest {
 
     @Test
     public void _不大于2公里时只收起步价6元() {
-        int amount = taxPayment.calc(1, 0);
+        long amount = taxPayment.calc(1, 0);
         assertThat(amount).isEqualTo(6);
     }
 
     @Test
     public void _超过2公里时每公里0点8元() {
-        int amount = taxPayment.calc(7, 0);
+        long amount = taxPayment.calc(7, 0);
         assertThat(amount).isEqualTo(10);
     }
 
     @Test
     public void _超过8公里则每公里加收百分之50长途费() {
-        int amount = taxPayment.calc(7, 0);
+        long amount = taxPayment.calc(7, 0);
         assertThat(amount).isEqualTo(10);
     }
 }
